refactor(chat): move chatPreview from widgets to chat/components
