+++ src/widgets/index.ts
@@ -1,2 +1,1 @@
-export * from './ChatPreview';
 export * from './EditWindow';
